+++ lib/osm.rb
@@ -139,7 +139,10 @@ module Osm
     }
 
     return permissions.inject({}) do |new_hash, (key, value)|
-      new_hash[key.to_sym] = (permissions_map[value.to_i] || [])
+      if ["badge", "member", "user", "register", "contact", "programme","events", "flexi", "finance"].include?(key)
+        # This is a permission we care about
+        new_hash[key.to_sym] = permissions_map[value.to_i]
+      end
       new_hash
     end
   end