+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 0.6.2
+
+  * Fix items not relevant to API appearing in permissions for a Section
+  * Fix started and started? methods for staged badges which have been completly done
+
 ## Version 0.6.1
 
   * Fixes to adventure badges